extract QuotedString
